added blader level increase on bey launch

+++ src/main/java/com/grillo78/BeyCraft/capabilities/BladerLevel.java
@@ -2,15 +2,15 @@ package com.grillo78.BeyCraft.capabilities;
 
 public class BladerLevel implements IBladerLevel{
 
-	private int bladerLevel=1;
+	private float bladerLevel=1;
 	
 	@Override
-	public void setBladerLevel(int level) {
+	public void setBladerLevel(float level) {
 		bladerLevel = level;
 	}
 
 	@Override
-	public int getBladerLevel() {
+	public float getBladerLevel() {
 		return bladerLevel;
 	}
